pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mops"]

[project]
name = "mops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.9.0,<2.10.0",
    "torchvision>=0.24.0,<0.25.0",
    "pytorch-lightning>=2.0.0",
    # Diffusion
    "diffusers>=0.36.0.dev0",
    "transformers>=4.57.3",
    "accelerate>=1.12.0",
    # Config
    "hydra-core>=1.3.0",
    "hydra-zen>=0.12.0",
    # Utils
    "einops>=0.8.1",
    "timm>=1.0.22",
    "torchdiffeq>=0.2.5",
    # Logging & Metrics
    "wandb",
    "tensorboard>=2.20.0",
    "torchmetrics>=1.8.2",
    # Data
    "datasets>=4.4.2",
    "h5py",
    "pandas>=2.3.3",
    # Typing
    "jaxtyping",
    "beartype",
    # Viz (optional)
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "scikit-learn>=1.8.0",
]

[dependency-groups]
dev = ["ruff", "pytest", "ipykernel>=7.1.0"]

# PyTorch sources
# Configured for Ubuntu with CUDA 12.8
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
diffusers = { git = "https://github.com/huggingface/diffusers" }
clip = { git = "https://github.com/openai/CLIP.git" }

[tool.ruff]
src = ["src"]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
# Ignore jaxtyping annotation errors (F722: forward annotation syntax, F821: undefined name)
ignore = ["F722", "F821", "UP037"]

[tool.uv]
link-mode = "symlink"
